Style: profile & search page
