+++ src/Components/Pages/Search Page/searchPage.module.css
@@ -1,0 +1,86 @@
+.searchPageContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  /* background: red; */
+}
+.searchbarWrapper {
+  /* width: 10rem; */
+  background: var(--background-white);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 2rem;
+  box-shadow: 0px 7px 12px -3px rgba(0, 0, 0, 0.151);
+  border-radius: 20px;
+}
+.searchBox {
+  height: 2rem;
+  border: none;
+  padding-left: 0.8rem;
+  border-radius: 20px 0 0 20px;
+}
+.searchBtn {
+  height: 2rem;
+  width: 3rem;
+  border: none;
+  border-radius: 0 20px 20px 0px;
+  font-size: 0.9rem;
+  color: var(--font-Color-White);
+  background: var(--primary-Col);
+  cursor: pointer;
+}
+.searchSuggestion {
+  font-size: 0.8rem;
+  margin-top: 1.5rem;
+  color: var(--font-Gray);
+  font-weight: 500;
+}
+.searchResultsWrapper {
+  margin-top: 2.5rem;
+  width: 70%;
+  text-align: center;
+}
+/* **** SEARCH CARD **** */
+
+.searchCard {
+  width: 100%;
+  /* background: red; */
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  box-shadow: -1px 1px 45px -21px rgba(32, 32, 32, 0.486);
+  padding: 0.5rem 0.3rem 0.5rem 0.3rem;
+  border-radius: 20px;
+}
+.searchCardAvatarWrapper {
+  /* background: blue; */
+  width: 4rem;
+}
+.searchCardAvatar {
+  width: 100%;
+  border-radius: 50px;
+  border: 2px solid white;
+  background: var(--avatar-Backdrop);
+}
+.searchCardUserInfo {
+  text-align: left;
+  margin-left: 1rem;
+}
+.searchCardFullname {
+  font-size: 1rem;
+  font-weight: 600;
+}
+.searchCardUsername {
+  font-size: 0.8rem;
+  font-weight: 500;
+  color: var(--font-Gray);
+}
+.searchCardBio {
+  width: 8rem;
+  font-size: 0.8rem;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
