+++ src/Components/index.js
@@ -2,9 +2,9 @@ import Homepage from "./Pages/Homepage";
 import SignIn from "./Pages/SignIn SignUp/SignIn";
 import SignUp from "./Pages/SignIn SignUp/SignUp";
 import NotFound from "./Pages/NotFound";
-import ProfilePage from "./Pages/ProfilePage";
+import ProfilePage from "./Pages/Profile Page/ProfilePage";
 import NotificationPage from "./Pages/NotificationPage";
-import SearchPage from "./Pages/SearchPage";
+import SearchPage from "./Pages/Search Page/SearchPage";
 import PostPage from "./Pages/PostPage";
 import IndexPage from "./Pages/Index Page/IndexPage";
